+++ libs/coin-modules/coin-evm/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @ledgerhq/coin-evm
 
+## 2.9.0
+
+### Minor Changes
+
+- [#8909](https://github.com/LedgerHQ/ledger-live/pull/8909) [`7224a7d`](https://github.com/LedgerHQ/ledger-live/commit/7224a7d3e5643fee75575fab74c72a2aca260be6) Thanks [@qperrot](https://github.com/qperrot)! - Fix bot device actions for polygon
+
+- [#8899](https://github.com/LedgerHQ/ledger-live/pull/8899) [`c62cec9`](https://github.com/LedgerHQ/ledger-live/commit/c62cec9a911f41d2383effeb2ea1f92756ce0a09) Thanks [@lambertkevin](https://github.com/lambertkevin)! - Prevent request on blockscout for ERC1155 NFTs as unsupported yet and add a `none` option for explorer's config to allow for node-only synchronization (will make token accounts disappear though)
+
+- [#8914](https://github.com/LedgerHQ/ledger-live/pull/8914) [`537141a`](https://github.com/LedgerHQ/ledger-live/commit/537141ab549b8dab57d3eb117e875faa67b54f4b) Thanks [@hedi-edelbloute](https://github.com/hedi-edelbloute)! - Bot feature : Add filter for features
+
+### Patch Changes
+
+- Updated dependencies [[`3a65633`](https://github.com/LedgerHQ/ledger-live/commit/3a6563309c8cacbd6e9a73e3044b1ff7c3966f87), [`537141a`](https://github.com/LedgerHQ/ledger-live/commit/537141ab549b8dab57d3eb117e875faa67b54f4b)]:
+  - @ledgerhq/coin-framework@0.23.0
+  - @ledgerhq/domain-service@1.2.17
+  - @ledgerhq/evm-tools@1.3.1
+
+## 2.9.0-next.0
+
+### Minor Changes
+
+- [#8909](https://github.com/LedgerHQ/ledger-live/pull/8909) [`7224a7d`](https://github.com/LedgerHQ/ledger-live/commit/7224a7d3e5643fee75575fab74c72a2aca260be6) Thanks [@qperrot](https://github.com/qperrot)! - Fix bot device actions for polygon
+
+- [#8899](https://github.com/LedgerHQ/ledger-live/pull/8899) [`c62cec9`](https://github.com/LedgerHQ/ledger-live/commit/c62cec9a911f41d2383effeb2ea1f92756ce0a09) Thanks [@lambertkevin](https://github.com/lambertkevin)! - Prevent request on blockscout for ERC1155 NFTs as unsupported yet and add a `none` option for explorer's config to allow for node-only synchronization (will make token accounts disappear though)
+
+- [#8914](https://github.com/LedgerHQ/ledger-live/pull/8914) [`537141a`](https://github.com/LedgerHQ/ledger-live/commit/537141ab549b8dab57d3eb117e875faa67b54f4b) Thanks [@hedi-edelbloute](https://github.com/hedi-edelbloute)! - Bot feature : Add filter for features
+
+### Patch Changes
+
+- Updated dependencies [[`3a65633`](https://github.com/LedgerHQ/ledger-live/commit/3a6563309c8cacbd6e9a73e3044b1ff7c3966f87), [`537141a`](https://github.com/LedgerHQ/ledger-live/commit/537141ab549b8dab57d3eb117e875faa67b54f4b)]:
+  - @ledgerhq/coin-framework@0.23.0-next.0
+  - @ledgerhq/domain-service@1.2.17-next.0
+  - @ledgerhq/evm-tools@1.3.1
+
 ## 2.8.1
 
 ### Patch Changes